Cloud book index change

+++ app/controllers/cloud_books_controller.rb
@@ -10,7 +10,6 @@ class CloudBooksController < ApplicationController
       options = { :email => current_user.email, :api_key => current_user.api_token }
 
 	@nodes = FindNodesByEmail.perform(options)
-puts @nodes.inspect
     if @nodes.class == Megam::Error
       #@res_msg="Sorry Something Wrong. Please contact #{ActionController::Base.helpers.link_to 'Our Support !.', "http://support.megam.co/"}."
     redirect_to new_cloud_book_path, :gflash => { :warning => { :value => "#{@cloud_books.some_msg[:msg]}", :sticky => false, :nodom_wrap => true } }
@@ -48,9 +47,7 @@ puts @nodes.inspect
       add_breadcrumb "Cloud_book's Definition", definition_path(params['format'])
       @cloud_books = current_user.cloud_books
 	logger.debug "Cloud Book Definition ==> "
-	puts params['format']
     @book = CloudBook.find(params['format'])
-	puts @book.inspect
   end
 
   def new
@@ -81,8 +78,6 @@ puts @nodes.inspect
     else
     #if @predef_cloud.some_msg[:msg_type] != "error"
       pred = FindPredefsByName.perform(predef_options)
-	puts "=================================> Predef COLLECTION <=============================================="
-	puts pred.inspect
       if pred.class == Megam::Error
         redirect_to new_cloud_book_path, :gflash => { :warning => { :value => "#{pred.some_msg[:msg]}", :sticky => false, :nodom_wrap => true } }
       else
@@ -102,7 +97,6 @@ puts @nodes.inspect
   def create
     options = { :email => current_user.email, :api_key => current_user.api_token, :node => mk_node(params, "server", "create") }
     @node = CreateNodes.perform(options)
-	puts "==================> @NODE in CB CREATE == > #{@node.inspect}"
     if @node.class == Megam::Error
       @res_msg="Sorry Something Wrong. Please contact #{ActionController::Base.helpers.link_to 'Our Support !.', "http://support.megam.co/"}."
       respond_to do |format|
@@ -129,12 +123,11 @@ puts @nodes.inspect
   end
 
   def show
-puts params
     @book = CloudBook.find(params[:id])
 	get_node = { :email => current_user.email, :api_key => current_user.api_token, :node => "#{params[:name]}" }
 	@node = FindNodeByName.perform(get_node)
 	logger.debug "Get Node By Name NODE ==> "
-
+	@requests = GetRequestsByNode.perform(get_node)
     if @node.class == Megam::Error
       @res_msg="Sorry Something Wrong. Please contact #{ActionController::Base.helpers.link_to 'Our Support !.', "http://support.megam.co/"}."
       respond_to do |format|
@@ -145,11 +138,9 @@ puts params
     #redirect_to new_cloud_book_path, :gflash => { :warning => { :value => "#{@node.some_msg[:msg]}", :sticky => false, :nodom_wrap => true } }
     else
 	@cloud_book = @node.lookup("#{params[:name]}")
-	puts "SHOW CLOUD BOOK ==> "
-	puts @cloud_book.inspect
       respond_to do |format|
         format.js {
-          respond_with(@cloud_book, :layout => !request.xhr? )
+          respond_with(@cloud_book, @requests, :layout => !request.xhr? )
         }
       end
     end
@@ -177,7 +168,6 @@ puts params
 
 n_name = params[:name]
 n_name = n_name[/[^.]+/]
-puts n_name
 comm = "#{@cloud_instruction.command}"
 comm["<node_name>"]="#{n_name}"
 
@@ -219,7 +209,6 @@ comm["<node_name>"]="#{n_name}"
 
     options = { :email => current_user.email, :api_key => current_user.api_token, :req => node_hash }
     @node = CreateRequests.perform(options)
-	puts "==================> @NODE in CB CREATE == > #{@node.inspect}"
     if @node.class == Megam::Error
       @res_msg="Sorry Something Wrong. Please contact #{ActionController::Base.helpers.link_to 'Our Support !.', "http://support.megam.co/"}."
       respond_to do |format|
@@ -228,10 +217,15 @@ comm["<node_name>"]="#{n_name}"
         }
       end
     else
+a = params[:n_hash]
+count = a["#{@book.name}"].count
+puts "TEST DELETE COUNT ==> #{count}"
+if count<2
 	@book.cloud_books_histories.each do |cbh|
 		cbh.destroy
 	end
 	@book.destroy
+end
 	redirect_to cloud_books_path, :gflash => { :success => { :value => "Cloud book and its histories are deleted successfully", :sticky => false, :nodom_wrap => true } }
 
     end
